Reindented and made minor changes.
Might still try to maintain this but for now, definitely using Python.

+++ src/fs.c
@@ -8,29 +8,29 @@ int exists(char *path) { return access(path, F_OK); }
 
 char *read_file(const char *path)
 {
-	FILE *file = fopen(path, "r");
-	if (!file) raise(1, "failed to open file: '%s'.", path);
+  FILE *file = fopen(path, "r");
+  if (!file) raise(1, "failed to open file: '%s'.", path);
 
-	/* Find size of file. */
-	fseek(file, 0, SEEK_END);
-	long file_size = ftell(file);
-	rewind(file);
+  // Find size of file.
+  fseek(file, 0, SEEK_END);
+  long file_size = ftell(file);
+  rewind(file);
 
-	/* Allocate memory for string. */
-	char *content = (char*)malloc((file_size+1) * sizeof(char));
-	if (!content) raise(1, "failed to allocate memory for file content.");
+  // Allocate memory for string.
+  char *content = (char*)malloc((file_size+1) * sizeof(char));
+  if (!content) raise(1, "failed to allocate memory for file content.");
 
-	/* Read file content into string. */
-	size_t read_size = fread(content, sizeof(char), file_size, file);
-	if ((int)read_size != (int)file_size)
-	{
-		fclose(file);
-		free(content);
-		raise(1, "failed to read '%s'.", path);
-	}
+  // Read file content into string.
+  size_t read_size = fread(content, sizeof(char), file_size, file);
+  if (read_size != file_size)
+  {
+    fclose(file);
+    free(content);
+    raise(1, "failed to read '%s'.", path);
+  }
 
-	content[file_size] = '\0';
-	fclose(file);
+  content[file_size] = '\0';
+  fclose(file);
 
-	return content;
+  return content;
 }
